Fix bandit camp fires
It was possible to provoke campfires on the edge of
the map by building roads over bandit campfires. This
resulted in a crash on diplomacy change because of
a blocking update.

This change saves the position and does not use the
entity to get the position as it was done before. So
campfires will not be created at the edge of the map.